automatically add undefined variables to the list
fixes #565

When an undefined variable is referred to anywhere in the question, it
is added to the "ungrouped variables" list.

I think this makes writing new questions easier: you can write out
the content areas and part settings as if variables exist, and then go
to the variables tab to define them.

This will also make it clear when there's a reference to an accidentally
undefined variable, due to renaming or typos.