starting work on queues
An ItemQueue is attached to a Project. It has a description,
instructions, a checklist template and a list of entries.

Queue entries contain a reference to an editor item (question or exam),
and have a note, a timeline of comments, and ticks against each of the
items in the queue's checklist.

At the moment there are views to create queues and entries, and to
manage the checklist and add comments to an individual entry.

Queues are listed at the top of the project's index page.

There are no permission checks on any of the views to do with queues at
the moment, and no views to edit or delete a queue or entry.

Because queues will need to give individual users access, I've got rid
of all the old *Access models and made a generic IndividualAccess model.
I've quickly checked that granting access to everything that had
individual access before still works, but need to have a more thorough
look.

I've also simplified the email notifications - there's just one class,
and it picks which templates to use based on the model names of the
objects involved.

There are migrations to add the new models and remove the old access
models, but there is not yet a data migration to create new instances of
IndividualAccess for each of the old *Access objects. The new migrations
need to be consolidated and this data migration added.

see #628