load extension scripts only when selected
Previously, scripts for all available extensions were included on
loading the question editor. This is slow and could cause errors!

Instead, scripts are only added to the page the first time the extension
is turned on.

Forthermore, all .js files in the extension's directory are included,
like for the real runtime.