show some project content on the index page
fixes #662

the project index page now has a panel in the main area showing the
top-level folders, and up to 3 of the most recently-changed items in the
project, along with links to browse all of the content.

The queues are moved to the sidebar, and the description is under the
content panel.

This also changes Project.members so they're sorted by last name.