always show the "upload a part type" button
The buttons to create or upload new CPTs were duplicated between two
cases in the template, but I forgot to include the upload button in the
case where there are no CPTs already.

This moves the buttons outside that whole if statement, so the same bit
produces the buttons in both cases.

fixes #686