pressing enter when in maths adds a <br>
fixes #371

I _think_ that this is robust: in TinyMCE, we intercept the 'keydown'
event to catch presses of the 'Enter' key. Writemaths now adds an
'in-maths' class to the element it's acting on, when the cursor is
inside maths. If that class is applied, then cancel the keypress and add
a <br> tag instead, then move the caret down.