URL_PREFIX, and get the CSRF token better
While setting up a second instance of the editor on a subpath,
I needed to set a global URL prefix, and to change the name of the
session and CSRF cookies.

This adds the URL_PREFIX setting to all URLs, and changes the code to
get the CSRF token in JS, to read from an input tag inserted by the
template engine's csrf_token command, rather than by looking at the
cookie.