Use django 4.1
A couple of required packages haven't been updated on PyPI to work with 4.1 yet,
so they're changed to be installed from git instead.

django-html-sanitizer seems to be unmaintained: it might be worth just
folding in the bits that we use.

I also took the opportunity to change the URL configs to use path()
instead of re_path()