more changes to the part editor
the "add a alternative/gap/step" buttons are now gathered together on
one line.

When dragging a part to rearrange it, it isn't collapsed any more. This
avoids a distracting flicker when just clicking on a part to select it.

The "add a part" modal is gone, and now all part types are shown on the
main screen, with a search box. Built-in and in-use part types are at
the top of the list and highlighted. Built-in types have descriptions
and help links.

Might have to look at this again if there end up being lots of published
custom part types.