tidying up the first_setup script
It's moved back to first_setup.py - now there's only the web version.

It does validation on the fields, and shows errors under the relevant
field.

The script to migrate the database and run other management commands is
now itself a management command, `first_setup_db`. This allows it to
avoid doing all the django setup stuff - manage.py does that for you.

Questions specific to the databases running on a server are hidden when
the DB engine is sqlite3.

The finish page shows the usernames of all superuser accounts.

I ran pylint over first_setup.py and first_setup_db.py.

I looked at having the web server quit once the files have been written,
but it needs to hang around to respond to static file requests.