show the full value of a variable under its definition
fixes #493 and #737

The "variable value" column in the list of variables now truncates its
display using the CSS `text-overflow: ellipsis` property, so text that
is too long will always have a visible ellipsis at the end.

In order to see the full code or display of a variable's value, this
commit adds a "generated value" box below the current variable's
definition and description fields.

This either contains the full JME representation of the value, or its
nice display, as appropriate.
An added benefit of this approach is that you can interact with HTML
elements.