in the extension part type's Testing tab, render the prompt
fixes #776.

An unintuitive side-effect of rendering the prompt in the testing tab is
that any HTML-valued variables, such as JSXGraph diagrams, disappear
from the preview in the Variables tab because they've been moved to the
part's Testing tab.

It might make sense to use a restored copy of the question's variables
in the part marking script, if that doesn't make everything too slow.