Adaptive marking: parts are referred to by unique ID
fixes #787

Each part object is given a unique ID on creation. The adaptive marking
variable replacement refers to the part to use by its ID, so it can be
unambiguously found after the order of parts changes.

I also had to add a rateLimit to the viewModel.allParts computed
observable so that it isn't recomputed during a drag-and-drop move,
causing the <select> box for the replacement part to lose the option for
that part and change the replacement value to whatever's left.