Editable package delete view: sidebar shows the context
issue #747 asked for the sidebar to be hidden on the delete file view.

I tried that, and decided that it would be helpful to show the context
for the file being deleted. (and it means there isn't a special case for
    the layout)

This commit rejigs the editable package views so that the delete file
view gets the necessary template context variables to show the sidebar
properly.