question editor: if generating variables didn't finish last time the page was open, don't autocalculate
When the question editor loads, it sets a local storage key. Once a set
of variables has been generated, that key is removed.

If the key is still set when the page loads, then variable generation
mustn't have finished last time the editor was open at this question, so
don't automatically generate variable values.

This gives authors an opportunity to fix questions with infinite loops
somewhere in a variable's definition.

This doesn't completely solve #713, but makes it possible to get
yourself out of the situation where a set of variables can't be
generated.