starting to rejig the exam options
The settings around when different kinds of feedback are shown are now
in a table, in the feedback tab. Everything else is in the navigation
tab, grouped under three headers: "Before the exam starts", "During the
exam", and "After the exam ends".

The editor interface is set up but the view model, loading and saving
code haven't been updated.