Give IDs to form controls in the editors
Until now, almost all of the form controls in the dynamic editors were
children of a <label> tag, so they'd pick up the text content of the
label without an ID attribute.

This had the side effect of making the entire width of the editor be the
clickable area for the control. For checkboxes, this meant that all of
the white space to the right of the checkbox was clickable.

In this commit, the *property template tags take an 'id' attribute, or
use the property name to calculate an automatic value, and the <label>
and <input> tags are separate.

fixes #828