add built-in constants before custom ones, and constants don't clash with auto-added variables
The built-in constants are added to the question's scope before any
custom ones, so disabling a built-in constant doesn't delete a defined
custom one with the same name.

When a variable with the same name as a custom constant is auto-added,
it doesn't cause a duplicate name error on the constant. This can happen
immediately after adding a custom constant with a name which has already
been used by reference.