can change resource filenames; improve editoritem uploads
Resource objects have a separate `filename` field which can be anything
you like, instead of being exactly the filename that the server uses on
disk.
This lets you keep the filename for a resource constant after
re-uploading it.
Need to add a warning for repeated filenames.

You can now upload a zip file in the upload editor item form. Either the
standalone or SCORM zip files produced by Numbas are fine.
They'll take the resources from the zip and attach them to the
right questions.

If you upload a single question (the zip file name starts with
'question-' and there's only one question in the .exam file) then only
a NewQuestion object will be created, instead of an exam containing that
question.

This also fixes a bug where all the questions in an uploaded exam would
have all the extensions used in the exam turned on, instead of just the
extensions required by that question.