Change RestrictAccessMixin to use exceptions instead of a boolean
I wanted to be able to give a message explaining why you don't have
access.

So now RestrictAccessMixin.can_access is check_access, and raises an
instance of editor.views.generic.NoAccessError when the user doesn't
have access. The error message can be displayed to the user.

The project no access template uses this to differentiate between cases
where you can't edit because you're not a member, and when you can't
edit because you only have view access.