make some DB queries more efficient
The queries to build the home page timeline, and to get the lists of
extensions and custom part types in the question editor, were doing
outer joins.

The problem is this pattern:

    Model.objects.filter(a) | related_model.models.all()

This produces an outer join, which is slow.

Instead, these queries could be written as:

    Model.objects.filter(Q(a) | pk__in=related_model.models.all())

This does an inner join, which is fast.