The preview view uses the generic runtime
Still need to deal with extensions, question resources, and locales.

The idea is that it's wasteful to compile the whole Numbas runtime for
every item that's previewed.

Instead, we could use the generic runtime and feed in the exam source.

It doesn't pass in the URLs of extensions or question resources, and I
need to make it take the locale as a URL query parameter.