add JS and CSS to extensions in the generic runtime
When a preview is displayed, the directories of all extensions are
walked to find .js and .css files for the runtime to load.
Hopefully this is fast enough that we can avoid caching this.

This also changes the `should_compile` logic to look at the runtime
directory in the compiler, and recompile if that's changed.

The preferred locale is passed to the Numbas runtime.