serve question resources in the preview view
This changes the URLs for the preview iframe.

The fundamental problem is that Numbas exams expect to load question
resources from the relative URL `resources/question-resources/...`.

So when using the generic runtime, the page's URL can't just be the
generic, static one. It has to be specific to the thing being previewed.

This adds a `(question|exam)_preview_file` route which will serve files
from a compiled preview. HTML files are served by Django. Any static
file in the generic compilation output will be served from the
MEDIA_URL. The redirect response adds a Last-Modified header so the
browser can use its cache.

Any requests under the `resources/question-resources` URL will try to
find the resource, and then redirect to the corresponding static URL
under MEDIA_URL, again with caching.