Question editor: getVariable copes with destructuring syntax
Apparently nobody else uses both destructuring and variable groups apart from me.

When loading the question and putting variables into their groups, it uses `Question.getVariable(name)` to get the Variable object.
But when the variable is defined by destructuring, `name` is a list of names.

This changes `getVariable` to split `name` on commas, and use just the first name from that list.