improve writemaths inside \begin{...} environments
The `findMaths` function in writemaths didn't cope with the case where the selection is just after a <br> element, where the Selection object's anchorNode is the <br> element, and the anchorIndex counts child nodes of the parent, instead of a position in the anchor node's text content.
This fixes that.

This also replaces the handler for linebreaks inside maths environments, to just simulate a shift+enter keydown event instead of trying to insert a <br> explicitly. TinyMCE does something to make sure the cursor ends up in the right place and the <br> isn't lost, but I don't know what it is.