JME parts: function sets and notation settings
Mathematical expression parts now have settings "notation" and "allowed functions".

The "notation" option determines the notation used to parse and display the answer, as well as variable value generator expressions.

The list of allowed functions can be used to restrict the functions or operations that the student can use in their answer.